requirements.txt
pydantic-ai
python-chess
PyQt5
logfire
rich